third_party/blink/web_tests/http/tests/misc/client-hints-accept-iframe.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
    var t = async_test("Verify that hints were not sent on iframe subresources");
    window.addEventListener("message", t.step_func(function (message) {
        console.log(message.data);
        assert_equals(message.data, "success");
        t.done();
    }));
</script>
<iframe src="resources/iframe-no-accept-ch.php" onerror="console.log('failed to load iframe');t.done()"></iframe>
